requirements.txt
pytorch>=1.11.0
tqdm>=4.64.0
torchvision>=0.12.0
torchmetrics>=0.11.2
numpy>=1.21.6
matplotlib>=3.5.1
tensorboardX>=2.5
tensorboard>=2.9.0
xarray>=2022.3.0
dask>=2022.7.0
netcdf4>=1.5.8
setuptools==59.5.0
xesmf>=0.6.2
cartopy>=0.20.2
numba>=0.55.1
